CHE-383: Move FqnProvider to che-core-ide-api
