+++ core/ide/che-core-ide-api/src/main/java/org/eclipse/che/ide/api/reference/FqnProvider.java
@@ -8,10 +8,16 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package org.eclipse.che.ide.reference;
+package org.eclipse.che.ide.api.reference;
 
 /**
- * The class provides methods which allows extract fqn from nodes which contains it.
+ * The class provides methods which allows to extract fully qualified name for nodes.
+ * Fully qualified name (Fqn) is an unambiguous name that specifies which object, function, or variable a call refers to
+ * without regard to the context of the call. In a hierarchical structure, a name is fully qualified when it "is complete in the
+ * sense that it includes
+ * (a) all names in the hierarchic sequence above the given element and
+ * (b) the name of the given element itself."
+ * Thus fully qualified names explicitly refer to namespaces that would otherwise be implicit because of the scope of the call.
  *
  * @author Dmitry Shnurenko
  */
