+++ core/ide/che-core-ide-app/src/test/java/org/eclipse/che/ide/reference/ShowReferencePresenterTest.java
@@ -14,6 +14,7 @@ import org.eclipse.che.api.workspace.shared.dto.ProjectConfigDto;
 import org.eclipse.che.ide.api.app.AppContext;
 import org.eclipse.che.ide.api.app.CurrentProject;
 import org.eclipse.che.ide.api.project.node.HasStorablePath;
+import org.eclipse.che.ide.api.reference.FqnProvider;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
